+++ app/src/main/java/com/example/myapplication/debitcards/domain/api/DebitCardsRepository.kt
@@ -2,6 +2,7 @@ package com.example.myapplication.debitcards.domain.api
 
 import com.example.myapplication.core.domain.models.Card
 import com.example.myapplication.core.domain.models.Result
+import java.math.BigDecimal
 
 interface DebitCardsRepository {
     suspend fun createDebitCard(
@@ -10,5 +11,5 @@ interface DebitCardsRepository {
         userId: Long
     ): Result<Card>
 
-    suspend fun depositToDebitCard(cardId: Long, amount: Long): Result<Unit>
+    suspend fun depositToDebitCard(cardId: Long, amount: BigDecimal): Result<Unit>
 }
